+++ esbuild.config.mjs
@@ -17,6 +17,7 @@ esbuild.build({
   target: "es2018",
   logLevel: "info",
   sourcemap: prod ? false : "inline",
+  minify: prod,
   treeShaking: true,
   outfile: "main.js",
   platform: "node",